Got rid of all compiler warnings in SciLexer.
- Compiles cleanly with /W4 /WX
- CRT_SECURE warnings are just disabled by the project-wide define
- Modifications to the project (doesn't look for 64bit porting warnings + stuff relative to optimization)

+++ scintilla/win32/PlatWin.cxx
@@ -2056,7 +2056,9 @@ double ElapsedTime::Duration(bool reset) {
 		LARGE_INTEGER lBegin;
 		lBegin.HighPart = bigBit;
 		lBegin.LowPart = littleBit;
-		double elapsed = lEnd.QuadPart - lBegin.QuadPart;
+		// NPPSTART Joce 06/16/09 Scintilla_clean_precomp
+		double elapsed = (double)(lEnd.QuadPart - lBegin.QuadPart);
+		// NPPEND
 		result = elapsed / static_cast<double>(frequency.QuadPart);
 	} else {
 		endBigBit = clock();
